News aggregator/requirements.txt
newspaper3k
nltk
bs4
requests
textblob